% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analysis}
\alias{analysis}
\title{Model analysis}
\usage{
analysis(results, ...)
}
\arguments{
\item{results}{Output of a model run}

\item{...}{Additional arguments}
}
\description{
Take the output of a model run and
return cost effectiveness values
}
