% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CombinedModel}
\alias{CombinedModel}
\title{Combine all submodels in to a single list}
\usage{
CombinedModel(...)
}
\arguments{
\item{...}{Submodels to combine}
}
\description{
Combine all submodels in to a single list
}
