% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DecisionTree.default}
\alias{DecisionTree.default}
\title{Default DecisionTree constructor}
\usage{
\method{DecisionTree}{default}(data, N = NA, ...)
}
\arguments{
\item{data}{Dataframe containing decision tree structure}

\item{N}{Total sample size}

\item{...}{Additional arguments}
}
\description{
Default DecisionTree constructor
}
