% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DecisionTree}
\alias{DecisionTree}
\title{Create a DecisionTree model}
\usage{
DecisionTree(data, ...)
}
\arguments{
\item{data}{Dataframe containing decision tree structure}

\item{...}{Additional arguments}
}
\description{
Create a DecisionTree model
}
