% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_costs.CombinedModel}
\alias{get_costs.CombinedModel}
\title{Extract costs from CombinedModel results}
\usage{
\method{get_costs}{CombinedModel}(results, ...)
}
\arguments{
\item{results}{CombinedModel output}

\item{...}{Additional arguments}
}
\description{
Extract costs from CombinedModel results
}
