% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_costs.DecisionTree}
\alias{get_costs.DecisionTree}
\title{Extract costs from DecisionTree results}
\usage{
\method{get_costs}{DecisionTree}(results, ...)
}
\arguments{
\item{results}{DecisionTree model output}

\item{...}{Additional arguments}
}
\description{
Extract costs from DecisionTree results
}
