% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_costs.default}
\alias{get_costs.default}
\title{Default get_costs method}
\usage{
\method{get_costs}{default}(results, ...)
}
\arguments{
\item{results}{Model output results}

\item{...}{Additional arguments}
}
\description{
Default get_costs method
}
