% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_costs.MarkovModel}
\alias{get_costs.MarkovModel}
\title{Extract costs from MarkovModel results}
\usage{
\method{get_costs}{MarkovModel}(results, ...)
}
\arguments{
\item{results}{MarkovModel output}

\item{...}{Additional arguments}
}
\description{
Extract costs from MarkovModel results
}
