% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_costs}
\alias{get_costs}
\title{Extract costs from model results}
\usage{
get_costs(results, ...)
}
\arguments{
\item{results}{Model output results}

\item{...}{Additional arguments passed to methods}
}
\description{
Extract costs from model results
}
