% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_effects.CombinedModel}
\alias{get_effects.CombinedModel}
\title{Extract health effects from CombinedModel results}
\usage{
\method{get_effects}{CombinedModel}(results, ...)
}
\arguments{
\item{results}{CombinedModel output}

\item{...}{Additional arguments}
}
\description{
Extract health effects from CombinedModel results
}
