% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_effects.DecisionTree}
\alias{get_effects.DecisionTree}
\title{Extract health effects from DecisionTree results}
\usage{
\method{get_effects}{DecisionTree}(results, ...)
}
\arguments{
\item{results}{DecisionTree model output}

\item{...}{Additional arguments}
}
\description{
Extract health effects from DecisionTree results
}
