% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_effects.default}
\alias{get_effects.default}
\title{Default get_effects method}
\usage{
\method{get_effects}{default}(results, ...)
}
\arguments{
\item{results}{Model output results}

\item{...}{Additional arguments}
}
\description{
Default get_effects method
}
