% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_effects.MarkovModel}
\alias{get_effects.MarkovModel}
\title{Extract health effects from MarkovModel results}
\usage{
\method{get_effects}{MarkovModel}(results, ...)
}
\arguments{
\item{results}{MarkovModel output}

\item{...}{Additional arguments}
}
\description{
Extract health effects from MarkovModel results
}
