% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_effects}
\alias{get_effects}
\title{Extract health effects (QALYs) from model results}
\usage{
get_effects(results, ...)
}
\arguments{
\item{results}{Model output results}

\item{...}{Additional arguments passed to methods}
}
\description{
Extract health effects (QALYs) from model results
}
