% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_decision_to_markov}
\alias{map_decision_to_markov}
\title{States from decision tree to Markov model}
\usage{
map_decision_to_markov(decision_result, mapping)
}
\value{
dataframe of probabilities by treatment
}
\description{
States from decision tree to Markov model
}
