% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_markov_to_decision}
\alias{map_markov_to_decision}
\title{States from Markov model to decision tree}
\usage{
map_markov_to_decision(dt_model, markov_result)
}
\description{
States from Markov model to decision tree
}
