% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_model.R
\name{markov_model}
\alias{markov_model}
\title{Native Markov Model calculation}
\usage{
markov_model(
  start_pop,
  p_matrix,
  state_c_matrix,
  state_q_matrix,
  n_cycles = 2,
  s_names = NULL,
  t_names = NULL
)
}
\arguments{
\item{start_pop}{Starting population array/matrix}

\item{p_matrix}{Transition probability matrix (or 3D array)}

\item{state_c_matrix}{State cost matrix}

\item{state_q_matrix}{State utility (QALY) matrix}

\item{n_cycles}{Number of cycles to run}

\item{s_names}{State names}

\item{t_names}{Treatment names}
}
\description{
Native Markov Model calculation
}
