% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MarkovModel.default}
\alias{MarkovModel.default}
\title{Default MarkovModel constructor}
\usage{
\method{MarkovModel}{default}(
  model = NA,
  init_probs = NA,
  trans_matrix = NA,
  cost_matrix = NA,
  q_matrix = NA,
  mapping = NA,
  N = NA,
  n_cycles = 2,
  ...
)
}
\arguments{
\item{model}{Optional existing model}

\item{init_probs}{Initial state probability array}

\item{trans_matrix}{Transition probability array}

\item{cost_matrix}{Cost array}

\item{q_matrix}{Quality of life (QALY) matrix}

\item{mapping}{Mapping vector from decision tree nodes to Markov states}

\item{N}{Total sample size}

\item{n_cycles}{Number of Markov cycles}

\item{...}{Additional arguments}
}
\description{
Default MarkovModel constructor
}
