% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MarkovModel}
\alias{MarkovModel}
\title{Create a MarkovModel model}
\usage{
MarkovModel(model, ...)
}
\arguments{
\item{model}{Optional existing model}

\item{...}{Additional arguments}
}
\description{
Create a MarkovModel model
}
