% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_model_diagram.R
\name{plot_decision_tree_diagram}
\alias{plot_decision_tree_diagram}
\title{Generate Decision Tree Diagram}
\usage{
plot_decision_tree_diagram(
  p_healthy = 0.8,
  p_preclin = 0.15,
  p_disease = 0.05,
  file = NULL
)
}
\arguments{
\item{p_healthy}{Initial probability of starting healthy}

\item{p_preclin}{Initial probability of starting pre-clinical}

\item{p_disease}{Initial probability of starting disease}

\item{file}{Optional filepath to save output plot (e.g. "plots/decision_tree_model.png")}
}
\description{
Generate Decision Tree Diagram
}
