% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_model.CombinedModel}
\alias{run_model.CombinedModel}
\title{Loop through each sequential submodel}
\usage{
\method{run_model}{CombinedModel}(model, ...)
}
\arguments{
\item{model}{CombinedModel object (list of submodels)}

\item{...}{Additional arguments}
}
\description{
Loop through each sequential submodel
}
