% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_model.DecisionTree}
\alias{run_model.DecisionTree}
\title{Run a DecisionTree model}
\usage{
\method{run_model}{DecisionTree}(model, ...)
}
\arguments{
\item{model}{DecisionTree model}

\item{...}{Additional arguments}
}
\description{
Run a DecisionTree model
}
