% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptors_markov_model.R
\name{run_model.MarkovModel}
\alias{run_model.MarkovModel}
\title{Native S3 method to run MarkovModel}
\usage{
\method{run_model}{MarkovModel}(model, ...)
}
\arguments{
\item{model}{MarkovModel object}

\item{...}{Additional arguments}
}
\description{
Native S3 method to run MarkovModel
}
