% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heemod_markov_model.R
\name{run_model_heemod}
\alias{run_model_heemod}
\title{Wrapper adaptor for heemod package (optional)}
\usage{
run_model_heemod(model)
}
\arguments{
\item{model}{A MarkovModel object}
}
\description{
Wrapper adaptor for heemod package (optional)
}
