% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_model.default}
\alias{update_model.default}
\title{Default update method}
\usage{
\method{update_model}{default}(model, result)
}
\arguments{
\item{model}{Model object}

\item{result}{Result object}
}
\description{
Default update method
}
