% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_model}
\alias{update_model}
\title{Provide additional components needed for each model}
\usage{
update_model(model, result)
}
\arguments{
\item{model}{Model to update}

\item{result}{Output from previous model}
}
\description{
Provide additional components needed for each model
}
