% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_model.MarkovModel}
\alias{run_model.MarkovModel}
\title{#' dummy function
#' @export
run_model.MarkovModel <- function(model) {

init_probs <- model$init_probs

res <- data.frame(
decision = c("Treatment A", "Treatment B"),
expected_cost = c(100, 100),
expected_eff = c(1,1))

if (!is.na(model$N)) {
res$expected_cost <- res$expected_cost * model$N
res$expected_eff <- res$expected_eff * model$N
}

structure(res,
class = c("output", class(model)))
}
Wrapper for adapter function}
\usage{
\method{run_model}{MarkovModel}(model)
}
\description{
#' dummy function
#' @export
run_model.MarkovModel <- function(model) {

init_probs <- model$init_probs

res <- data.frame(
decision = c("Treatment A", "Treatment B"),
expected_cost = c(100, 100),
expected_eff = c(1,1))

if (!is.na(model$N)) {
res$expected_cost <- res$expected_cost * model$N
res$expected_eff <- res$expected_eff * model$N
}

structure(res,
class = c("output", class(model)))
}
Wrapper for adapter function
}
