% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_model.DecisionTree}
\alias{update_model.DecisionTree}
\title{Decision tree update method}
\usage{
\method{update_model}{DecisionTree}(model, result)
}
\arguments{
\item{model}{DecisionTree object}

\item{result}{Output object from previous model}
}
\description{
Decision tree update method
}
