% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_model.MarkovModel}
\alias{update_model.MarkovModel}
\title{Markov model update method}
\usage{
\method{update_model}{MarkovModel}(model, result)
}
\arguments{
\item{model}{MarkovModel object}

\item{result}{Output object from previous model}
}
\description{
Markov model update method
}
